refactored orders panel
